requirements.txt
fastapi
sqlmodel
uvicorn
python-jose
passlib[bcrypt]
python-dotenv
asyncpg

